requirements.txt
twelvedata>=1.2.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
streamlit>=1.29.0
python-dotenv>=1.0.0
